RpcFullBackEnd/service2/requirements.txt
nameko==2.13.0
psycopg2-binary==2.9.3
redis==4.2.2
